styling and cleanup
